Move `impl` blocks out of `rustc_middle/src/mir/syntax.rs`.
As the comment at the top says, this file is not supposed to contain any
code. But some has crept in. This commit moves it out.
